refactor: added function docstrings and refactored into multiple files
