+++ .env
@@ -1,0 +1,2 @@
+SERP_API_KEY="<your-serp-api-key>"
+OPENAI_API_KEY="<your-openai-api-key>"
